Add `count` and `enumerate` functions
`count` counts the number of occurences a predicate is true for each
element of a `Foldable`.

`enumerate` pairs each element of a list or vector with its index
for `List` the index is a `Nat`, for `Vect` the index is a `Fin`
of the length of the vector in order to allow indexing into it.